move files to hopefully fix dependency resolution
